+++ CHANGELOG.md
@@ -13,6 +13,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   in the `AggregateManager` the `EventStore` associated type as `Box<dyn EventStore<Manager = >>` (with `Send` + `Sync` 
   bounds).
 
+# Changed
+
+- [[#117]]: 
+  - `AggregateState::new` second parameter from `Uuid` to `impl Into<Uuid>`.
+  - `AggregateManager::load` first parameter from `Uuid` to `impl Into<Uuid>`.
+  - `AggregateState::delete` first parameter from `Uuid` to `impl Into<Uuid>`.
+
 ---
 
 ## [0.7.0]
@@ -125,6 +132,7 @@ Refer to: [#107], [#108] and [#109]
 [0.7.0]: https://github.com/primait/event_sourcing.rs/compare/0.6.2...0.7.0
 [0.6.2]: https://github.com/primait/event_sourcing.rs/compare/0.6.1...0.6.2
 
+[#117]: https://github.com/primait/event_sourcing.rs/pull/117
 [#114]: https://github.com/primait/event_sourcing.rs/pull/114
 [#109]: https://github.com/primait/event_sourcing.rs/pull/109
 [#108]: https://github.com/primait/event_sourcing.rs/pull/108